requirements.txt
asttokens==3.0.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
comtypes==1.4.10
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
dnspython==2.7.0
eventlet==0.39.1
executing==2.2.0
Flask==3.1.0
flask-cors==5.0.1
Flask-MySQL==1.6.0
Flask-SocketIO==5.5.1
fonttools==4.57.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mysql-connector-python==9.2.0
nest-asyncio==1.6.0
nltk==3.8.1
numpy==1.26.4
packaging==24.2
pandas==2.2.1
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.7
prompt_toolkit==3.0.50
psutil==7.0.0
pure_eval==0.2.3
PyAudio==0.2.14
Pygments==2.19.1
PyMySQL==1.1.1
pyparsing==3.2.3
pypiwin32==223
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.0
python-socketio==5.13.0
pyttsx3==2.98
pytz==2025.2
pywin32==310
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
scikit-learn==1.4.2
scipy==1.15.2
seaborn==0.13.2
simple-websocket==1.1.0
six==1.17.0
SpeechRecognition==3.10.0
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
wsproto==1.2.0



